agent/pyproject.toml
[project]
name = "graph-analysis-agent"
version = "0.1.0"
description = "LangGraph-based agent for graph theory analysis with Gemini integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "networkx>=3.4.0",
    "matplotlib>=3.9.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "pillow>=11.0.0",
    "numpy>=2.1.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "poethepoet>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "api", "tools", "config.py", "state.py", "graph_parsers.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]


[tool.poe]
executor.type = "uv"

[tool.poe.tasks]
start = "uvicorn api.main:app --reload"